Combine duplicated MockPOSCatalogPersistenceService
