+++ Modules/Tests/YosemiteTests/Tools/POS/POSCatalogFullSyncServiceTests.swift
@@ -170,22 +170,3 @@ struct POSCatalogFullSyncServiceTests {
     }
 }
 
-// MARK: - Mock POSCatalogPersistenceService
-
-private final class MockPOSCatalogPersistenceService: POSCatalogPersistenceServiceProtocol {
-    private(set) var replaceAllCatalogDataCallCount = 0
-    private(set) var lastPersistedCatalog: POSCatalog?
-    private(set) var lastPersistedSiteID: Int64?
-
-    func replaceAllCatalogData(_ catalog: POSCatalog, siteID: Int64) async throws {
-        replaceAllCatalogDataCallCount += 1
-        lastPersistedSiteID = siteID
-        lastPersistedCatalog = catalog
-    }
-
-    func persistIncrementalCatalogData(_ catalog: POSCatalog, siteID: Int64) async throws {}
-
-    func loadSite(siteID: Int64) async throws -> POSSite? { nil }
-
-    func updateSite(_ site: POSSite) async throws {}
-}
