+++ Modules/Tests/YosemiteTests/Tools/POS/POSCatalogIncrementalSyncServiceTests.swift
@@ -174,33 +174,3 @@ struct POSCatalogIncrementalSyncServiceTests {
         #expect(site2ModifiedAfter == lastFullSyncDate)
     }
 }
-
-// MARK: - Mock Classes
-
-private final class MockPOSCatalogPersistenceService: POSCatalogPersistenceServiceProtocol {
-    private(set) var persistIncrementalCatalogDataCallCount = 0
-    private(set) var persistIncrementalCatalogDataLastPersistedCatalog: POSCatalog?
-    private(set) var persistIncrementalCatalogDataLastPersistedSiteID: Int64?
-    var persistIncrementalCatalogDataError: Error?
-
-    private var storedSites: [Int64: POSSite] = [:]
-
-    func replaceAllCatalogData(_ catalog: POSCatalog, siteID: Int64) async throws {}
-
-    func persistIncrementalCatalogData(_ catalog: POSCatalog, siteID: Int64) async throws {
-        persistIncrementalCatalogDataCallCount += 1
-        persistIncrementalCatalogDataLastPersistedSiteID = siteID
-        persistIncrementalCatalogDataLastPersistedCatalog = catalog
-        if let error = persistIncrementalCatalogDataError {
-            throw error
-        }
-    }
-
-    func loadSite(siteID: Int64) async throws -> POSSite? {
-        storedSites[siteID]
-    }
-
-    func updateSite(_ site: POSSite) async throws {
-        storedSites[site.siteID] = site
-    }
-}
